Move exception into exception namespace
